FEAT: Number of pizza's created will always cover background
FEAT: Number of pizza's created will always cover background.

I added an eventlistener. When more pizzas are needed to cover the
background, all current moving pizzas are removed, then new pizzas are
created.

KNOW ISSUE: between the width: 1265px -> 1000px the pizzas on each row
line up with pizzas on the corresponding column. This is a Math issue in
how the offset left is created but I'm not sure how to fix it.